Move SpeakerPlayItem to MusicPlayItem
