+++ dirigera-client-api/src/main/java/de/dvdgeisler/iot/dirigera/client/api/model/device/speaker/SpeakerAudioAttributes.java
@@ -2,19 +2,20 @@ package de.dvdgeisler.iot.dirigera.client.api.model.device.speaker;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import de.dvdgeisler.iot.dirigera.client.api.model.music.MusicPlayItem;
 
 @JsonInclude(Include.NON_NULL)
 public class SpeakerAudioAttributes {
   public String serviceType;
   public String providerType;
-  public SpeakerPlayItem playItem;
-  public SpeakerPlayItem nextPlayItem;
+  public MusicPlayItem playItem;
+  public MusicPlayItem nextPlayItem;
 
   public SpeakerAudioAttributes() {
   }
 
-  public SpeakerAudioAttributes(String serviceType, String providerType, SpeakerPlayItem playItem,
-      SpeakerPlayItem nextPlayItem) {
+  public SpeakerAudioAttributes(String serviceType, String providerType, MusicPlayItem playItem,
+      MusicPlayItem nextPlayItem) {
     this.serviceType = serviceType;
     this.providerType = providerType;
     this.playItem = playItem;
