+++ dirigera-client-api/src/main/java/de/dvdgeisler/iot/dirigera/client/api/model/music/MusicPlayList.java
@@ -1,15 +1,13 @@
 package de.dvdgeisler.iot.dirigera.client.api.model.music;
 
-import de.dvdgeisler.iot.dirigera.client.api.model.device.speaker.SpeakerPlayItem;
-
 import java.util.List;
 
 public class MusicPlayList {
     public String id;
-    public List<SpeakerPlayItem> playItems;
+    public List<MusicPlayItem> playItems;
     public String title;
 
-    public MusicPlayList(final String id, final List<SpeakerPlayItem> playItems, final String title) {
+    public MusicPlayList(final String id, final List<MusicPlayItem> playItems, final String title) {
         this.id = id;
         this.playItems = playItems;
         this.title = title;
